Small fix and prevent downloading json if exist

+++ formicID/AntWeb/json_to_csv.py
@@ -108,7 +108,8 @@ def batch_json_to_csv(
     csvname,
     dataset,
     quality='low',
-    output_dir=None
+    output_dir=None,
+    overwrite=False
 ):
     """From a json file or batch of json files a csvfile is created with
     relevant information for downloading the images and naming the files.
@@ -116,9 +117,16 @@ def batch_json_to_csv(
     Args:
         csvname (str): Name of the output csv file.
         dataset (str): The name of the dataset (directory).
+        quality (str): Set the image quality when downloading the images. Each
+            different image quality has its own unique link, that is why it
+            needs to be set when filtering the JSON file. Set the quility to
+            either one of (from highest quality to lowest): `high`,
+            `thumbview`, `medium`, `low`. Defaults to `low`.
         output_dir (str): name of the output directory. Defaults to `None`. If
             `None`, the images will be placed in an `image` subfolder in the
             dataset directory.
+        overwrite (bool): Whether to overwrite the csv file with (new) json
+            data, if the csv file already exists.
 
     Returns:
         file: A csv file containing the necesarry information for the scrape
@@ -139,34 +147,43 @@ def batch_json_to_csv(
     else:
         outputdir = os.mkdir(os.path.join('data',
                                           output_dir))
-    nb_files = len(os.listdir(input_direc))
-    if nb_files == 0:
-        raise AssertionError('There are no files in the input directory.')
-    df2 = pd.DataFrame()
-    columns = ['catalog_number',
-               'scientific_name',
-               'shot_type',
-               'image_url']
-    for filename in tqdm(os.listdir(input_direc),
-                         desc='Converting JSON files to csv',
-                         total=nb_files,
-                         unit='JSON files'):
-        if filename.endswith('.json'):
-            with open(os.path.join(input_direc,
-                                   filename)) as data_file:
-                lst = _filter_json(data_file, quality=quality)
-                df = pd.DataFrame(lst,
-                                  columns=columns)
-                df2 = df2.append(df)
-    # replace spaces between genus and species names with underscores
-    df2.replace('\s+', '_',
-                regex=True,
-                inplace=True)
-    df2.columns = columns
-    df2.to_csv(os.path.join(output_dir,
-                            csvname),
-               index=False,
-               header=True)
+    if not os.path.isfile(os.path.join(output_dir, csvname)):
+        pass
+    if os.path.isfile(os.path.join(output_dir, csvname)):
+        if overwrite == False:
+            logging.info('Csv file already exists, and will not be '
+                         'overwritten, due to `overwrite` is set to `False`.')
+        else:
+            pass
+    else:
+        nb_files = len(os.listdir(input_direc))
+        if nb_files == 0:
+            raise AssertionError('There are no files in the input directory.')
+        df2 = pd.DataFrame()
+        columns = ['catalog_number',
+                   'scientific_name',
+                   'shot_type',
+                   'image_url']
+        for filename in tqdm(os.listdir(input_direc),
+                             desc='Converting JSON files to csv',
+                             total=nb_files,
+                             unit='JSON files'):
+            if filename.endswith('.json'):
+                with open(os.path.join(input_direc,
+                                       filename)) as data_file:
+                    lst = _filter_json(data_file, quality=quality)
+                    df = pd.DataFrame(lst,
+                                      columns=columns)
+                    df2 = df2.append(df)
+        # replace spaces between genus and species names with underscores
+        df2.replace('\s+', '_',
+                    regex=True,
+                    inplace=True)
+        df2.columns = columns
+        df2.to_csv(os.path.join(output_dir,
+                                csvname),
+                   index=False,
+                   header=True)
     logging.info('All JSON files are read, filtered and added to the csv '
                  'file. "{0}" was created in {1}'.format(csvname,
                                                          output_dir))
